+++ modules.tf
@@ -1,6 +1,6 @@
 module "support_subnet" {
   source  = "claranet/subnet/azurerm"
-  version = "5.0.0"
+  version = "6.0.0"
 
   environment    = var.environment
   location_short = var.location_short
@@ -14,9 +14,14 @@ module "support_subnet" {
   custom_subnet_name = var.custom_bastion_subnet_name
   subnet_cidr_list   = var.subnet_cidr_list
 
-  route_table_rg    = var.route_table_rg
-  route_table_name  = var.route_table_name
-  service_endpoints = var.service_endpoints
+  route_table_rg   = var.route_table_rg
+  route_table_name = var.route_table_name
+
+  service_endpoints           = var.service_endpoints
+  service_endpoint_policy_ids = var.service_endpoint_policy_ids
+
+  private_link_endpoint_enabled = var.private_link_endpoint_enabled
+  private_link_service_enabled  = var.private_link_service_enabled
 }
 
 resource "azurerm_subnet_network_security_group_association" "subnet_bastion_association" {
@@ -26,7 +31,7 @@ resource "azurerm_subnet_network_security_group_association" "subnet_bastion_ass
 
 module "support_nsg" {
   source  = "claranet/nsg/azurerm"
-  version = "5.1.0"
+  version = "7.0.0"
 
   client_name         = var.client_name
   environment         = var.environment
@@ -41,6 +46,24 @@ module "support_nsg" {
   use_caf_naming = var.use_caf_naming
 
   custom_network_security_group_name = var.custom_security_group_name
+
+  # Custom
+  additional_rules = var.additional_rules
+
+  # Flow Logs
+  flow_log_enabled                               = var.flow_log_enabled
+  flow_log_logging_enabled                       = var.flow_log_logging_enabled
+  network_watcher_name                           = var.network_watcher_name
+  network_watcher_resource_group_name            = var.network_watcher_resource_group_name
+  flow_log_storage_account_id                    = var.flow_log_storage_account_id
+  flow_log_retention_policy_enabled              = var.flow_log_retention_policy_enabled
+  flow_log_retention_policy_days                 = var.flow_log_retention_policy_days
+  flow_log_traffic_analytics_enabled             = var.flow_log_traffic_analytics_enabled
+  log_analytics_workspace_guid                   = var.log_analytics_workspace_guid
+  log_analytics_workspace_location               = var.log_analytics_workspace_location
+  log_analytics_workspace_id                     = var.log_analytics_workspace_id
+  flow_log_traffic_analytics_interval_in_minutes = var.flow_log_traffic_analytics_interval_in_minutes
+  flow_log_location                              = var.flow_log_location
 
   # Tags
   default_tags_enabled = var.default_tags_enabled
@@ -65,7 +88,7 @@ resource "azurerm_network_security_rule" "ssh_rule" {
 }
 
 module "bastion" {
-  source = "github.com/claranet/terraform-azurerm-bastion-vm.git?ref=v6.0.0"
+  source = "github.com/claranet/terraform-azurerm-bastion-vm.git?ref=v7.0.0"
 
   client_name         = var.client_name
   location            = var.location
@@ -82,6 +105,8 @@ module "bastion" {
   # Custom bastion VM
   custom_vm_name     = var.custom_vm_name
   custom_vm_hostname = var.custom_vm_hostname
+  backup_policy_id   = var.backup_policy_id
+  identity           = var.identity
 
   # VM Network
   subnet_bastion_id  = module.support_subnet.subnet_id
@@ -94,10 +119,11 @@ module "bastion" {
   ssh_private_key = var.ssh_private_key
 
   # VM OS Image ref
-  storage_image_publisher = var.storage_image_publisher
-  storage_image_offer     = var.storage_image_offer
-  storage_image_sku       = var.storage_image_sku
-  storage_image_id        = var.storage_image_id
+  storage_image_publisher      = var.storage_image_publisher
+  storage_image_offer          = var.storage_image_offer
+  storage_image_sku            = var.storage_image_sku
+  storage_image_id             = var.storage_image_id
+  storage_os_disk_account_type = var.storage_os_disk_account_type
 
   # VM OS Disk params
   storage_os_disk_custom_name = var.storage_os_disk_custom_name
@@ -130,4 +156,5 @@ module "bastion" {
   pubip_extra_tags                = var.pubip_extra_tags
   storage_os_disk_tagging_enabled = var.storage_os_disk_tagging_enabled
   storage_os_disk_extra_tags      = var.storage_os_disk_extra_tags
+  extensions_extra_tags           = var.extensions_extra_tags
 }