+++ variables-logs.tf
@@ -16,12 +16,6 @@ variable "use_legacy_monitoring_agent" {
   default     = false
 }
 
-variable "log_analytics_workspace_guid" {
-  description = "GUID of the Log Analytics Workspace to link with"
-  type        = string
-  default     = null
-}
-
 variable "log_analytics_workspace_key" {
   description = "Access key of the Log Analytics Workspace to link with"
   type        = string